// frontend/src/app/model/answer.ts
/**
 * H2MS API
 * API for interacting with the H2MS backend.
 *
 * OpenAPI spec version: 1.0
 *
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

import { Event } from './event';
import { Question } from './question';


export interface Answer {
    answerType?: string;
    event?: Event;
    id?: number;
    question?: string;
    value?: string;
}